moved Upload compoent code to App.js file
