--- src/Components/Upload/styles.css
@@ -1,30 +1,0 @@
-.droparea {
-  width: 240px;
-  height: 240px;
-  align-items: center;
-  text-align: center;
-  border: 2px dashed grey;
-}
-.button {
-  border-radius: 50%;
-  background-color: rgb(241, 22, 22);
-  color: white;
-  text-align: center;
-  font-size: x-large;
-  position: absolute;
-  padding: 5px;
-  top: 50%;
-  left: 50%;
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-.imageContainer {
-  box-sizing: border-box;
-}
-.row {
-  display: flex;
-}
-.column {
-  flex: 100%;
-  padding: 5px;
-}
